Remove SetupHelper class
